extract useIntersectionObserver
