+++ src/components/common/InfiniteScroll/index.tsx
@@ -1,28 +1,5 @@
-import { type MutableRefObject, useEffect, useRef, useState, type ReactElement, useCallback } from 'react'
-
-const useIntersectionObserver = (element: MutableRefObject<HTMLElement | null>): boolean => {
-  const [isIntersecting, setIsIntersecting] = useState<boolean>(false)
-  const observer = useRef<IntersectionObserver | undefined>()
-
-  const callback = useCallback(([entry]: IntersectionObserverEntry[]) => {
-    setIsIntersecting(entry.isIntersecting)
-  }, [])
-
-  useEffect(() => {
-    if (element.current) {
-      observer.current = new IntersectionObserver(callback)
-
-      observer.current.observe(element.current)
-    }
-
-    return () => {
-      observer.current?.disconnect()
-      observer.current = undefined
-    }
-  }, [callback, element])
-
-  return isIntersecting
-}
+import { useEffect, useRef, type ReactElement } from 'react'
+import useIntersectionObserver from '@/hooks/useIntersectionObserver'
 
 const InfiniteScroll = ({ onLoadMore }: { onLoadMore: () => void }): ReactElement => {
   const elementRef = useRef<HTMLDivElement | null>(null)
